langchain_debatesys/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
pydantic>=2.0.0
streamlit>=1.29.0
python-dotenv>=1.0.0
huggingface-hub>=0.20.2
numpy>=1.24.0